Changes
Hey so here are the changes that I made to the program. I made sure that
i commented on what i did and my initials next to it "MR". I am going to
make the addcontact function better, but i just wanted to upload it how
did it at first.

Also be sure to change the default of the console window, both widths to
100, so everything fits neatly in the window.