Remove function
So I wasn't able to find a proper function in the library to remove data
or delete it. So what i did instead was make the edit function replace
all the fields (with the exception of the id), with blanks. I'll
continue to work on it before the final submission.